refactor: Refatora cliente-criado-subscriber
