+++ cliente-criado-subscriber/tests/unit/test-handler.test.ts
@@ -1,18 +1,23 @@
 import { SNSEvent } from 'aws-lambda';
+import { CognitoIdentityProviderClient } from '@aws-sdk/client-cognito-identity-provider';
+import { mockClient } from 'aws-sdk-client-mock';
 import { lambdaHandler } from '../../app';
-import { expect, describe, it, jest, beforeEach } from '@jest/globals';
-import { AmazonCognitoService } from '../../cognito.service';
+import { expect, describe, it, beforeEach } from '@jest/globals';
 
+const cognitoMock = mockClient(CognitoIdentityProviderClient);
+
+/**
+ * To be sure that unit tests are independent from each other,
+ * reset mock behavior between the tests.
+ */
 beforeEach(() => {
-    jest.spyOn(AmazonCognitoService, 'criarUsuario').mockImplementation(() => Promise.resolve());
+    cognitoMock.reset();
 });
 
 describe('Unit test for app handler', function () {
     it('verifies successful response', async () => {
         const event: SNSEvent = JSON.parse("{\"Records\":[{\"EventSource\":\"aws:sns\",\"EventVersion\":\"1.0\",\"EventSubscriptionArn\":\"arn:aws:sns:us-east-1:900534935988:ClienteCriado:83a74a64-46e7-4e5a-827c-14edec9cbd52\",\"Sns\":{\"Type\":\"Notification\",\"MessageId\":\"81c44c07-4a0d-579e-9ff2-1263c84e8998\",\"TopicArn\":\"arn:aws:sns:us-east-1:900534935988:ClienteCriado\",\"Subject\":null,\"Message\":\"{\\\"_nome\\\":\\\"Fulano Da Silva Santos\\\",\\\"_email\\\":\\\"fulano@gmail.com\\\",\\\"_cpf\\\":\\\"74993021801\\\"}\",\"Timestamp\":\"2024-03-08T03:03:31.540Z\",\"SignatureVersion\":\"1\",\"Signature\":\"M2zwMeE4ncGC8aLa3/otY2Ni/P1mPIs3InFewxJQ9TB+s9fqYIoOsNuEcp+UMWzTqv8FFdiOy9DL4N+QT0Bh14rXOacSWz/POD6hV8YEKfE0tY/V9tBcGiwNj1movmwzSxQjA/MPhMNuakqIs6XlH8B5IS18oQsPSwWaLboPY+rhNVgDIGEmWl2iiB78cgALYSKzR7O4IYyFWOV9/KYc2tuL41J27/Ooy0g7P+uMWooUl1Hdq3YxRxKC/8Tcd4hXqCGuR9anp5ed1He5qdiAVjm87Zh5RahhaFeYXfvLnmsssdVBf/SPFNxHbQhKDSYw7EBe7Npi12+DR5d7bQD1/g==\",\"SigningCertUrl\":\"https://sns.us-east-1.amazonaws.com/SimpleNotificationService-60eadc530605d63b8e62a523676ef735.pem\",\"UnsubscribeUrl\":\"https://sns.us-east-1.amazonaws.com/?Action=Unsubscribe&SubscriptionArn=arn:aws:sns:us-east-1:900534935988:ClienteCriado:83a74a64-46e7-4e5a-827c-14edec9cbd52\",\"MessageAttributes\":{}}}]}");
 
-        await lambdaHandler(event);
-
-        expect(AmazonCognitoService.criarUsuario).toBeCalledTimes(1);
+        expect(() => lambdaHandler(event)).not.toThrow();
     });
 });
